t2129: refactor(email-pipeline): decompose email python cluster
Decompose the 5 email-pipeline Python files flagged by qlty file-complexity
into 11 new focused modules (16 files total). All existing CLI entry points,
import statements, and public APIs remain unchanged via re-export barrels.

Split summary:
- email_jmap_adapter.py (1705→244 lines): → _index, _transport, _helpers, _sync, _commands
- email-voice-miner.py (1045→430 lines): → voice_patterns, voice_analyzer
- extraction_pipeline.py (947→315 lines): → extraction_models, extraction_validate
- email-summary.py (614→287 lines): → email_summary_heuristic
- email_normaliser.py (488→153 lines): → email_normaliser_sections

Verification:
- python3 -m py_compile passes on all 16 files
- qlty smells --all reports 0 file-complexity on email/extraction cluster
- All CLI --help entry points respond correctly
- All import consumers (email-to-markdown.py, entity-extraction.py, etc.) compile

For #19222
Resolves #19225
